event-transformer: extract app selector
