+++ src/apps/selection/app.rs
@@ -131,17 +131,13 @@ impl App for Selection {
 #[cfg(test)]
 mod test {
     use crate::midi::{Error, Event, Image};
+    use crate::midi::features::{R, AppSelector};
     use crate::apps;
     use super::*;
 
     struct Transformer {}
-    impl EventTransformer for Transformer {
-        fn into_index(&self, _event: Event) -> Result<Option<u16>, Error> { Err(Error::Unsupported) }
-        fn into_app_index(&self, _event: Event) -> Result<Option<u16>, Error> { Err(Error::Unsupported) }
-        fn into_color_palette_index(&self, _event: Event) -> Result<Option<u16>, Error> { Err(Error::Unsupported) }
-        fn from_image(&self, _image: Image) -> Result<Event, Error> { Err(Error::Unsupported) }
-        fn from_index_to_highlight(&self, _index: u16) -> Result<Event, Error> { Err(Error::Unsupported) }
-        fn from_app_colors(&self, app_colors: Vec<[u8; 3]>) -> Result<Event, Error> {
+    impl AppSelector for Transformer {
+        fn from_app_colors(&self, app_colors: Vec<[u8; 3]>) -> R<Event> {
             let mut bytes = vec![];
             for app_color in &app_colors {
                 bytes.push(app_color[0]);
@@ -150,6 +146,12 @@ mod test {
             }
             return Ok(Event::SysEx(bytes));
         }
+    }
+    impl EventTransformer for Transformer {
+        fn into_index(&self, _event: Event) -> Result<Option<u16>, Error> { Err(Error::Unsupported) }
+        fn into_color_palette_index(&self, _event: Event) -> Result<Option<u16>, Error> { Err(Error::Unsupported) }
+        fn from_image(&self, _image: Image) -> Result<Event, Error> { Err(Error::Unsupported) }
+        fn from_index_to_highlight(&self, _index: u16) -> Result<Event, Error> { Err(Error::Unsupported) }
         fn from_color_palette(&self, _color_palette: Vec<[u8; 3]>) -> Result<Event, Error> { Err(Error::Unsupported) }
     }
