+++ src/midi/devices/launchpadpro/device.rs
@@ -46,10 +46,6 @@ impl EventTransformer for LaunchpadProEventTransformer {
         return super::index::into_index(event);
     }
 
-    fn into_app_index(&self, event: Event) -> Result<Option<u16>, Error> {
-        return super::index::into_app_index(event);
-    }
-
     fn into_color_palette_index(&self, event: Event) -> Result<Option<u16>, Error> {
         return super::index::into_color_palette_index(event);
     }
@@ -60,10 +56,6 @@ impl EventTransformer for LaunchpadProEventTransformer {
 
     fn from_index_to_highlight(&self, index: u16) -> Result<Event, Error> {
         return super::index::from_index_to_highlight(index);
-    }
-
-    fn from_app_colors(&self, app_colors: Vec<[u8; 3]>) -> Result<Event, Error> {
-        return super::index::from_app_colors(app_colors);
     }
 
     fn from_color_palette(&self, color_palette: Vec<[u8; 3]>) -> Result<Event, Error> {
