+++ src/midi/devices/launchpadpro/index.rs
@@ -19,25 +19,6 @@ pub fn into_index(event: Event) ->  Result<Option<u16>, Error> {
     });
 }
 
-pub fn into_app_index(event: Event) ->  Result<Option<u16>, Error> {
-    return Ok(match event {
-        // event must be a "note down" with a strictly positive velocity
-        Event::Midi([176, data1, data2, _]) if data2 > 0 => {
-            // the device provides a 10x10 grid if you count the buttons on the sides
-            let row = data1 / 10;
-            let column  = data1 % 10;
-
-            // we’ll use the last column on the right to select applications
-            if row >= 1 && row <= 8 && column == 9 {
-                Some(8 - row).map(|index| index.into())
-            } else {
-                None
-            }
-        },
-        _ => None,
-    });
-}
-
 pub fn into_color_palette_index(event: Event) ->  Result<Option<u16>, Error> {
     return Ok(match event {
         // event must be a "note down" with a strictly positive velocity
@@ -68,27 +49,6 @@ pub fn from_index_to_highlight(index: u16) -> Result<Event, Error> {
     let led = row * 10 + column;
 
     let bytes = vec![240, 0, 32, 41, 2, 16, 40, led, 45, 247];
-    return Ok(Event::SysEx(bytes));
-}
-
-pub fn from_app_colors(app_colors: Vec<[u8; 3]>) -> Result<Event, Error> {
-    if app_colors.len() > 8 {
-        return Err(Error::OutOfBoundIndexError);
-    }
-
-    let mut bytes = vec![240, 0, 32, 41, 2, 16, 11];
-
-    for index in 0..app_colors.len() {
-        let led = (89 - 10 * index) as u8;
-        bytes.append(&mut vec![
-            led,
-            app_colors[index][0] / 4,
-            app_colors[index][1] / 4,
-            app_colors[index][2] / 4,
-        ]);
-    }
-    bytes.push(247);
-
     return Ok(Event::SysEx(bytes));
 }
 
@@ -178,56 +138,6 @@ mod tests {
             08, 09, 10, 11, 12, 13, 14, 15,
             00, 01, 02, 03, 04, 05, 06, 07,
         ]
-            .iter()
-            .map(|index| Some(*index))
-            .collect::<Vec<Option<u16>>>();
-
-        assert_eq!(expected_output, actual_output);
-    }
-
-    #[test]
-    fn into_app_index_given_incorrect_status_should_return_none() {
-        let event = Event::Midi([128, 89, 10, 0]);
-        assert_eq!(None, into_app_index(event).expect("into_app_index should not fail"));
-    }
-
-    #[test]
-    fn into_app_index_given_low_velocity_should_return_none() {
-        let event = Event::Midi([176, 89, 0, 0]);
-        assert_eq!(None, into_app_index(event).expect("into_app_index should not fail"));
-    }
-
-    #[test]
-    fn into_app_index_given_out_of_grid_value_should_return_none() {
-        let events = vec![
-            [176, 08, 10, 0],
-            [176, 09, 10, 0],
-            [176, 18, 10, 0],
-            [176, 28, 10, 0],
-            [176, 38, 10, 0],
-            [176, 48, 10, 0],
-            [176, 58, 10, 0],
-            [176, 68, 10, 0],
-            [176, 78, 10, 0],
-            [176, 88, 10, 0],
-            [176, 98, 10, 0],
-            [176, 99, 10, 0],
-        ];
-
-        for event in events {
-            let event = Event::Midi(event);
-            assert_eq!(None, into_app_index(event).expect("into_app_index should not fail"));
-        }
-    }
-
-    #[test]
-    fn into_app_index_should_correct_value() {
-        let actual_output = vec![19, 29, 39, 49, 59, 69, 79, 89]
-            .iter()
-            .map(|code| into_app_index(Event::Midi([176, *code, 10, 0])).expect("into_app_index should not fail"))
-            .collect::<Vec<Option<u16>>>();
-
-        let expected_output = vec![7, 6, 5, 4, 3, 2, 1, 0]
             .iter()
             .map(|index| Some(*index))
             .collect::<Vec<Option<u16>>>();
